chore: add style in storybook template
